pinetime: Update the linker scripts to allow a larger payload
The adjusted linker scripts allow the reloader to overwrite its own
data structures. This works providing they are authored in chunks so
that the chunk that gets overwritten has already been written at its
final destination.

The factory firmware (which is pretty big) has been chunked accordingly
as a demonstration.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>