Add code to wake spinor.
Many PineTime payloads (currently including mcuboot) are broken if booted
with the spinor in deep powerdown. Let's wake the spinor during reloading
to avoid bugs when switching from a firmware with power down support to
one without.

The changes here are suprisingly large because we have to refactor the
existing spi code to untangle it from the st7789 driver.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>