Avoid automatic reboot on panic (if possible)
Currently when reloader panics is stops and waits for the watchdog to fire.
That's all very well... but if went to get a cup of tea when doing an OTA
update you are left with no clue that something bad has happened.

The new code only activiates on platforms with a physical button (since
we need a physical button to provoke a watchdog reset).

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>