update-reloader-factory
Update with a newer version of the InifiniTime bootloader (v1.0.1) since this
is needed for the reloader to properly support the BY25Q32 SPINOR in newer
PineTime devices.

The resulting wasp reloader_factory.zip was successfully tested on the Pinetime
with the new BY25Q32 memory chip. We also tested it on the original SPI chip
on the older Pinetime devices.

Fixes: https://github.com/wasp-os/wasp-os/issues/519
Signed-off-by: GaryM0101 <gary@geekslounge.com>
Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>